[edit] variable type]

+++ 14-binary_tree_balance.c
@@ -7,7 +7,7 @@
 */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	size_t l_height, r_height;
+	int l_height, r_height;
 
 	if (tree == NULL || (tree->left != NULL && tree->right != NULL))
 		return (0);
